% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_lcm.R
\name{frac_lcm}
\alias{frac_lcm}
\alias{frac_gcd}
\title{Least common multiple and greatest common divisor}
\usage{
frac_lcm(..., max = 1e+07)

frac_gcd(...)
}
\arguments{
\item{...}{Integer vectors or vectors that can be coerced to integer.}

\item{max}{If the least common multiple is greater than \code{max}, \code{max} is
returned instead.}
}
\value{
An integer.
}
\description{
Least common multiple and greatest common divisor
}
\examples{
frac_lcm(1, 2, 3, 4, 5, 6)
x <- 1:6
frac_lcm(x)
frac_lcm(x, 7)

frac_gcd(12, 42, 60)
y <- c(12, 42, 60)
frac_gcd(y)
frac_gcd(y, 39)
}
