% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_mat.R
\name{frac_mat}
\alias{frac_mat}
\alias{as.frac_mat}
\alias{is.frac_mat}
\title{Convert decimals to a matrix of numerators and denominators}
\usage{
frac_mat(
  x,
  base_10 = FALSE,
  common_denom = FALSE,
  mixed = FALSE,
  max_denom = 1e+07
)

as.frac_mat(x)

is.frac_mat(x)
}
\arguments{
\item{x}{A vector of decimals or, for \code{as.frac_mat()}, a character vector
created by \code{\link[=fracture]{fracture()}}}

\item{base_10}{If \code{TRUE}, all denominators will be a power of 10.}

\item{common_denom}{If \code{TRUE}, all fractions will have the same denominator.

If the least common denominator is greater than \code{max_denom},
\code{max_denom} is used.}

\item{mixed}{If \code{TRUE}, \code{integer} components will be displayed separately
from fractional components for \code{x} values greater than 1.

If \code{FALSE}, improper fractions will be used for \code{x} values greater than 1.}

\item{max_denom}{All denominators will be less than or equal to
\code{max_denom}.

If \code{base_10} is \code{TRUE}, the maximum denominator will be the largest power
of 10 less than \code{max_denom}.

A \code{max_denom} greater than the inverse square root of
\link[=.Machine]{machine double epsilon} will produce a warning because floating
point rounding errors can occur when denominators grow too large.}
}
\value{
A matrix with the same number of columns as the length of \code{x} and
rows for \code{integer}s (if \code{mixed} is \code{TRUE}), \code{numerator}s,
and \code{denominator}s.
}
\description{
Convert decimals to a matrix of numerators and denominators
}
\examples{
x <- (6:1) / (1:6)

frac_mat(x)
frac_mat(x, common_denom = TRUE)

frac_mat(x, base_10 = TRUE)
frac_mat(x, base_10 = TRUE, max_denom = 100)
frac_mat(x, base_10 = TRUE, common_denom = TRUE)
frac_mat(x, base_10 = TRUE, common_denom = TRUE, max_denom = 100)

frac_mat(x, mixed = TRUE)
frac_mat(x, mixed = TRUE, common_denom = TRUE)
frac_mat(x, mixed = TRUE, base_10 = TRUE)
frac_mat(x, mixed = TRUE, base_10 = TRUE, max_denom = 100)
frac_mat(x, mixed = TRUE, base_10 = TRUE, common_denom = TRUE)
frac_mat(x, mixed = TRUE, base_10 = TRUE, common_denom = TRUE, max_denom = 100)
}
\seealso{
\code{\link[=fracture]{fracture()}} to return a character vector of fractions.
}
