% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracture-package.R
\docType{package}
\name{fracture-package}
\alias{fracture-package}
\alias{_PACKAGE}
\alias{reprex-package}
\title{fracture: Convert decimals to fractions}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions for converting decimals to a matrix of
  numerators and denominators or a character vector of fractions.
  Supports mixed or improper fractions, finding common denominators for vectors
  of fractions, limiting denominators to powers of ten, and limiting
  denominators to a maximum value.
  Also includes helper functions for finding the least common multiple and
  greatest common divisor for a vector of integers.
  Implemented using C++ for maximum speed.
}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
